requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
basicsr==1.4.2
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.16.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
dill==0.3.7
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
fasteners==0.14.1
filelock==3.13.1
filterpy==1.4.5
firebase-admin==6.3.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.2
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gyp==0.1
httplib2==0.20.2
huggingface-hub==0.20.2
HyperPyYAML==1.2.2
idna==3.3
imageio==2.33.1
importlib-metadata==7.0.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lmdb==1.4.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
pooch==1.8.0
prompt-toolkit==3.0.42
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.25
stack-data==0.6.3
streamlit==1.30.0
sympy==1.12
systemd-python==234
tb-nightly==2.16.0a20240112
tenacity==8.2.3
tensorboard-data-server==0.7.2
tf_keras-nightly==2.16.0.dev2024011210
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
usb-creator==0.3.7
validators==0.22.0
virtualenv==20.25.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
yt-dlp==2023.12.30
zipp==1.0.0
